fix test regressions

+++ python/tvm/relay/op/image/image.py
@@ -67,7 +67,7 @@ def resize(
         The resized result.
     """
     if isinstance(size, Constant):
-        size = list(size.data.asnumpy())
+        size = list(size.data.asnumpy().astype("int32"))
     if isinstance(size, Expr):
         return _dyn_make.resize(
             data, size, layout, method, coordinate_transformation_mode, out_dtype
